modify benchmark
